refactor: footer structure
Images downloaded
Change folder structure

Francisco Zambrano
