+++ src/assets/png/colombia-logo.png
