+++ src/components/Footer/Footer.scss
@@ -1,0 +1,58 @@
+.footer {
+  background-color: #ff9c00;
+  color: #fff;
+  padding: 3rem;
+  margin-top: 2em;
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  font-size: 1.1em;
+
+  &__logo {
+    width: auto;
+    height: 33px;
+    margin-left: 0.2rem;
+  }
+}
+
+.presidential-candidates {
+
+  width: 12%;
+
+  &__images {
+    display: flex;
+    justify-content: center;
+
+    img {
+      width: 120px;
+      height: auto;
+    }
+  }
+
+  &__information {
+    background-color: #c06db4;
+    border-radius: 1rem;
+    height: 50px;
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+
+    a {
+      text-decoration: none;
+      color: white;
+    }
+  }
+}
+
+.colombia-logo {
+  flex: 1 1 auto;
+  height: 250px;
+}
+
+.developers-information {
+  width: 20%;
+
+  &__logo {
+    @extend .footer__logo;
+  }
+}
